Add tab-completion file-name expansion.
Add a “-c” option which enables tab-completion for file-names.  With
it enabled, when entered text contains multiple words (for example
“soffice docs”) or contains a slash (for example “~/bin”), pressing
C-i or Tab will expand the last argument as a file name.

If there were multiple expansions the longest common prefix is
substituted.  Otherwise, if there was exactly one expansion a slash or
space is added ofter it depending on whether it expanded to
a directory name or not.

One known limitation is that if expanded file name contains
white-space, user will have to quote the argument herself or otherwise
when executing it will likely be interpreted as separate arguments.

This is based on 2009 patch by Jeremy Jay which can be found at
<http://lists.suckless.org/dwm/0901/7355.html>, but lacks buffer
overflow bug. ;)